refactor(backend): extract provider and Codex subdomains
